Add Castling
Castling is tricky because it requires two moves - the king
and the rook need to move, so we now return a list of moves
rather than a single move. In all other cases this will
just have one entry.

Changed the is_reachable function to enable the rook to move
through the king when the move category is castling (just
return true) and use the src_file and src_rank attributes
in Move to select the correc rook.