New move mechanism
Changed the way the source square is found. From the move notation
we know the destination square and the piece type. Therefore, we
can imagine a piece of that type on the destination square and see
if it can legally move to a square that contains that piece. If so,
we know that that square has a piece that can move to the
destination square.

Also required changes on how to check whether the king was in check.
Still need to implement how en passant works

Removed the Turn class as actually it makes more sense to hold
everything in the Move class